Tambah Util untuk error handling
